src/vlastudio/policy/qwen25dp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ilstd_qwen25"
version = "0.1.0"
description = "A Plug-and-Play Imitation-Learning Playground for Robotics"
readme = "README.md"
requires-python = "==3.10.*"
license = { text = "MIT" }
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "torchaudio>=2.4.0",
    "transformers==4.52.4",
    "accelerate",
    "peft==0.17.1",
    "opencv-python>=4.9.0.80",
    "rospkg==1.5.0",
    "imageio==2.37.0",
    "pillow>=11.3.0",
    "diffusers==0.35.1",
    "tensorflow==2.15.0",
    "tf-keras==2.15.0",
    "tensorflow_datasets",
    "tianshou==0.2.0",
    "pytest",
    "dm_control==1.0.14",
    "matplotlib",
    "robosuite==1.4.0",
    "imageio-ffmpeg==0.6.0",
    "pynput",
    "numpy==1.26.4",
    "mujoco==2.3.7",
    "dlimp @ git+https://github.com/kvablack/dlimp@5edaa4691567873d495633f2708982b42edf1972",
    "torchdata<=0.9.0",
    "lerobot",
    "qwen-vl-utils",
    "rich",
    "timm>=0.9.10",
    "draccus",
    "tensorflow_graphics",
    "loguru",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true